+++ jsonmodel/exceptions.py
@@ -1,3 +1,4 @@
+''' a package of exception classes '''
 __author__ = 'rcj1492'
 __created__ = '2016.01'
 __license__ = 'MIT'