+++ jsonmodel/loader.py
@@ -1,3 +1,4 @@
+''' a package of data loading functions '''
 __author__ = 'rcj1492'
 __created__ = '2016.01'
 __license__ = 'MIT'